Create AdidasUseCase.py
ADIDAS CASE STUDY

Load the data and print/make yourself familiar with the schema. Count the rows in the (raw) data set.
2. Apply some data profiling on the data set and suggest three data quality measures to improve the data set for the purpose of
this case study
3. Make sure your data set is cleaned, so we for example don't include in results with empty/null "titles" and "number of pages" is
greater than 20 and "publishing year" is after 1950. If you decided to add additional filters due to the data profiling exercise
above, pls. state them as well. Count the rows in the filtered dataset.
4. Run the following queries on the cleaned dataset:
▪ Get the first book / author which was published - and the last one.
▪ Find the top 5 genres with most published books.
▪ Retrieve the top 5 authors who (co-)authored the most books.
▪ Per publish year, get the number of authors that published at least one book
▪ Find the number of authors and number of books published per month for years between 1950 and 1970!